requirements.txt
langchain-groq
python-dotenv
pydantic
beautifulsoup4
lxml
requests
pathlib
pypdf
langchain
langchain-community
chromadb
pandas
pytest
pytest-mock
schedule
fastapi
uvicorn
httpx
streamlit
